% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/display-taylor.R
\name{display_taylor_all}
\alias{display_taylor_all}
\title{Display all color palettes}
\usage{
display_taylor_all(n = NULL, select = NULL, exact.n = TRUE)
}
\arguments{
\item{n}{The number of values to display from the palette}

\item{select}{A list of names of existing palettes to select}

\item{exact.n}{If TRUE, only display palettes with a color number given by n}
}
\value{
An image showing all palettes
}
\description{
Display all color palettes
}
