% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/display-taylor.R
\name{display_taylor_pal}
\alias{display_taylor_pal}
\title{Display the color palette}
\usage{
display_taylor_pal(name, n = 10)
}
\arguments{
\item{name}{The name of the palette}

\item{n}{The number of values to display from the palette}
}
\value{
An image showing the given palette
}
\description{
Display the color palette
}
