% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taylor_palette.R
\name{pal_pal}
\alias{pal_pal}
\title{Create a reversible, manual palette}
\usage{
pal_pal(values, direction = 1)
}
\arguments{
\item{values}{The vector of values for the colors in the palette}

\item{direction}{Indicates whether the palette should be reversed}
}
\value{
A function that creates a palette
}
\description{
Create a function for a palette that is reversible by the user. Heavily based
upon \code{\link[scales:manual_pal]{scales::manual_pal()}}, with influence from \code{\link[scales:brewer_pal]{scales::brewer_pal()}}.
}
