% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taylor_palette.R
\name{taylor_palette}
\alias{taylor_palette}
\title{Get a palette for Taylor Swift albums}
\usage{
taylor_palette(name, n, type = c("discrete", "continuous"))
}
\arguments{
\item{name}{The name of the desired palette.}

\item{n}{The number of items in the palette}

\item{type}{The type of the palette, either c("discrete", "continuous")}
}
\value{
The resulting palette.
}
\description{
Get a palette for Taylor Swift albums
}
