% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taylor_palette.R
\docType{data}
\name{taylor_palettes}
\alias{taylor_palettes}
\title{Complete list of palettes}
\format{
An object of class \code{list} of length 24.
}
\usage{
taylor_palettes
}
\description{
Use \code{\link{taylor_palette}} to construct palettes of desired length.
}
\keyword{datasets}
