% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taylor_palette.R
\name{taylor_palette}
\alias{taylor_palette}
\title{Get a palette for Taylor Swift albums}
\usage{
taylor_palette(name, n, type = c("discrete", "continuous"), direction = 1)
}
\arguments{
\item{palette}{The name of the desired palette.}
}
\value{
The resulting palette.
}
\description{
Get a palette for Taylor Swift albums
}
